+++ android/src/main/java/notification/listener/service/NotificationConstants.java
@@ -1,6 +1,7 @@
 package notification.listener.service;
 
 public final class NotificationConstants {
+    public static String ID = "notification_id";
     public static String INTENT = "slayer.notification.listener.service.intent";
     public static String PACKAGE_NAME = "package_name";
     public static String NOTIFICATION_CONTENT = "message";
@@ -9,4 +10,5 @@ public final class NotificationConstants {
     public static String EXTRAS_PICTURE = "extras_picture";
     public static String NOTIFICATIONS_ICON = "notifications_icon";
     public static String IS_REMOVED = "is_removed";
+    public static String CAN_REPLY = "can_reply_to_it";
 }