+++ android/src/main/java/notification/listener/service/models/ActionCache.java
@@ -1,0 +1,7 @@
+package notification.listener.service.models;
+
+import java.util.HashMap;
+
+abstract public class ActionCache {
+    public static HashMap<Integer, Action> cachedNotifications = new HashMap<>();
+}