+++ lib/notification_event.dart
@@ -1,6 +1,14 @@
 import 'dart:typed_data';
 
+import 'notification_listener_service.dart';
+
 class ServiceNotificationEvent {
+  /// the notification id
+  int? id;
+
+  /// check if we can reply the Notification
+  bool? canReply;
+
   /// if the notification has an extras image
   bool? hasExtrasPicture;
 
@@ -35,6 +43,8 @@ class ServiceNotificationEvent {
   String? content;
 
   ServiceNotificationEvent({
+    this.id,
+    this.canReply,
     this.hasExtrasPicture,
     this.hasRemoved,
     this.extrasPicture,
@@ -44,19 +54,37 @@ class ServiceNotificationEvent {
     this.content,
   });
 
-  ServiceNotificationEvent.fromMap(Map<dynamic, dynamic> json) {
-    hasExtrasPicture = json['hasExtrasPicture'];
-    hasRemoved = json['hasRemoved'];
-    extrasPicture = json['notificationExtrasPicture'];
-    packageName = json['packageName'];
-    title = json['title'];
-    notificationIcon = json['notificationIcon'];
-    content = json['content'];
+  ServiceNotificationEvent.fromMap(Map<dynamic, dynamic> map) {
+    id = map['id'];
+    canReply = map['canReply'];
+    hasExtrasPicture = map['hasExtrasPicture'];
+    hasRemoved = map['hasRemoved'];
+    extrasPicture = map['notificationExtrasPicture'];
+    packageName = map['packageName'];
+    title = map['title'];
+    notificationIcon = map['notificationIcon'];
+    content = map['content'];
+  }
+
+  /// send a direct message reply to the incoming notification
+  Future<bool> sendReply(String message) async {
+    if (!canReply!) throw Exception("The notification is not replyable");
+    try {
+      return await methodeChannel.invokeMethod<bool>("sendReply", {
+            'message': message,
+            'notificationId': id,
+          }) ??
+          false;
+    } catch (e) {
+      rethrow;
+    }
   }
 
   @override
   String toString() {
     return '''ServiceNotificationEvent(
+      id: $id
+      can reply: $canReply
       packageName: $packageName
       title: $title
       content: $content