+++ lib/notification_listener_service.dart
@@ -5,14 +5,13 @@ import 'dart:io';
 import 'package:flutter/services.dart';
 import 'package:notification_listener_service/notification_event.dart';
 
+const MethodChannel methodeChannel =
+    MethodChannel('x-slayer/notifications_channel');
+const EventChannel _eventChannel = EventChannel('x-slayer/notifications_event');
+Stream<ServiceNotificationEvent> _stream = const Stream.empty();
+
 class NotificationListenerService {
   NotificationListenerService._();
-
-  static const MethodChannel _methodeChannel =
-      MethodChannel('x-slayer/notifications_channel');
-  static const EventChannel _eventChannel =
-      EventChannel('x-slayer/notifications_event');
-  static Stream<ServiceNotificationEvent> _stream = const Stream.empty();
 
   /// stream the incoming Accessibility events
   static Stream<ServiceNotificationEvent> get notificationsStream {
@@ -30,17 +29,17 @@ class NotificationListenerService {
   /// it will open the notification settings page and return `true` once the permission granted.
   static Future<bool> requestPermission() async {
     try {
-      return await _methodeChannel.invokeMethod('requestPermission');
+      return await methodeChannel.invokeMethod('requestPermission');
     } on PlatformException catch (error) {
       log("$error");
       return Future.value(false);
     }
   }
 
-  /// check if notification permession is enebaled
+  /// check if notification permission is enebaled
   static Future<bool> isPermissionGranted() async {
     try {
-      return await _methodeChannel.invokeMethod('isPermissionGranted');
+      return await methodeChannel.invokeMethod('isPermissionGranted');
     } on PlatformException catch (error) {
       log("$error");
       return false;