+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 ## 0.0.2
 
 * usage improve
+## 0.0.3
+
+* check if you can reply to a notification 
+* add the possibility to reply to a notification